Move ResourceCalendar to its own component
Refs #565
