+++ app/pages/resource/reservation-calendar/ReservationCalendarContainer.spec.js
@@ -1,8 +1,6 @@
 import { expect } from 'chai';
 import MockDate from 'mockdate';
 import React from 'react';
-import DayPicker from 'react-day-picker';
-import MomentLocaleUtils from 'react-day-picker/moment';
 import { browserHistory } from 'react-router';
 import simple from 'simple-mock';
 
@@ -10,6 +8,7 @@ import DateHeader from 'shared/date-header';
 import ReservationCancelModal from 'shared/modals/reservation-cancel';
 import ReservationInfoModal from 'shared/modals/reservation-info';
 import ReservationSuccessModal from 'shared/modals/reservation-success';
+import ResourceCalendar from 'shared/resource-calendar';
 import Resource from 'utils/fixtures/Resource';
 import TimeSlot from 'utils/fixtures/TimeSlot';
 import { getResourcePageUrl } from 'utils/resourceUtils';
@@ -33,9 +32,7 @@ describe('pages/resource/reservation-calendar/ReservationCalendarContainer', () 
   const resource = Resource.build();
 
   const defaultProps = {
-    availability: {},
     actions,
-    currentLanguage: 'en',
     date: '2015-10-11',
     isAdmin: false,
     isEditing: false,
@@ -70,17 +67,8 @@ describe('pages/resource/reservation-calendar/ReservationCalendarContainer', () 
       wrapper = getWrapper(props);
     });
 
-    it('renders DayPicker', () => {
-      expect(wrapper.find(DayPicker).length).to.equal(1);
-    });
-
-    it('renders a calendar-legend with correct labels', () => {
-      expect(wrapper.find('.calendar-legend .free').text())
-        .to.equal('ReservationCalendarPickerLegend.free');
-      expect(wrapper.find('.calendar-legend .busy').text())
-        .to.equal('ReservationCalendarPickerLegend.busy');
-      expect(wrapper.find('.calendar-legend .booked').text())
-      .to.equal('ReservationCalendarPickerLegend.booked');
+    it('renders ResourceCalendar', () => {
+      expect(wrapper.find(ResourceCalendar).length).to.equal(1);
     });
 
     it('renders DateHeader', () => {
@@ -198,60 +186,6 @@ describe('pages/resource/reservation-calendar/ReservationCalendarContainer', () 
           };
           makeRenderTests(props, options);
         });
-      });
-    });
-  });
-
-  describe('DayPicker', () => {
-    let instance;
-    let dayWrapper;
-    before(() => {
-      const wrapper = getWrapper({
-        availability: {
-          '2015-10-01': { percentage: 0 },
-          '2015-10-02': { percentage: 50 },
-          '2015-10-03': { percentage: 81 },
-          '2015-10-04': { percentage: 100 },
-        },
-      });
-      instance = wrapper.instance();
-      dayWrapper = wrapper.find(DayPicker);
-    });
-
-    it('renders correct props', () => {
-      expect(dayWrapper.prop('disabledDays').before.valueOf()).to.be.closeTo(
-        new Date().valueOf(),
-        100000
-      );
-      expect(dayWrapper.prop('enableOutsideDays')).to.be.true;
-      expect(dayWrapper.prop('initialMonth')).to.deep.equal(new Date(defaultProps.date));
-      expect(dayWrapper.prop('locale')).to.equal('en');
-      expect(dayWrapper.prop('localeUtils')).to.equal(MomentLocaleUtils);
-      expect(dayWrapper.prop('onDayClick')).to.equal(instance.onDateChange);
-      expect(dayWrapper.prop('selectedDays')).to.deep.equal(new Date(defaultProps.date));
-    });
-
-    describe('modifiers', () => {
-      it('is available if percentage is greater than 80', () => {
-        const func = dayWrapper.prop('modifiers').available;
-        expect(func(new Date('2015-10-01'))).to.be.false;
-        expect(func(new Date('2015-10-02'))).to.be.false;
-        expect(func(new Date('2015-10-03'))).to.be.true;
-        expect(func(new Date('2015-10-04'))).to.be.true;
-      });
-      it('is busy if percentage is lower than 80', () => {
-        const func = dayWrapper.prop('modifiers').busy;
-        expect(func(new Date('2015-10-01'))).to.be.false;
-        expect(func(new Date('2015-10-02'))).to.be.true;
-        expect(func(new Date('2015-10-03'))).to.be.false;
-        expect(func(new Date('2015-10-04'))).to.be.false;
-      });
-      it('is booked if percentage is 0', () => {
-        const func = dayWrapper.prop('modifiers').booked;
-        expect(func(new Date('2015-10-01'))).to.be.true;
-        expect(func(new Date('2015-10-02'))).to.be.false;
-        expect(func(new Date('2015-10-03'))).to.be.false;
-        expect(func(new Date('2015-10-04'))).to.be.false;
       });
     });
   });
