+++ app/shared/resource-calendar/index.js
@@ -1,0 +1,3 @@
+import ResourceCalendar from './ResourceCalendar';
+
+export default ResourceCalendar;
