+++ app/shared/resource-calendar/resource-calendar.less
@@ -1,0 +1,97 @@
+.calendar-availability {
+  @available-color: #D1FAD1;
+  @busy-color: #F6EAAC;
+  @booked-color: #F5A3A3;
+  display: inline-block;
+
+  .DayPicker {
+    padding: 0;
+    border: 1px solid @gray-light;
+    border-radius: 5px;
+    overflow-x: hidden;
+  }
+
+  .DayPicker-NavBar {
+    padding: .75rem .5rem;
+  }
+
+  .DayPicker-Caption {
+    height: auto;
+    font-weight: 600;
+    line-height: 3rem;
+  }
+
+  .DayPicker-Month {
+    margin: 0;
+  }
+
+  .DayPicker-Day {
+    border: none;
+    padding: 1.5rem;
+    line-height: 1rem;
+    &--disabled {
+      cursor: pointer;
+    }
+    &--available:not(.DayPicker-Day--disabled):not(.DayPicker-Day--selected) {
+      background-color: @available-color;
+    }
+    &--busy:not(.DayPicker-Day--disabled):not(.DayPicker-Day--selected)  {
+      background-color: @busy-color;
+    }
+    &--booked:not(.DayPicker-Day--disabled):not(.DayPicker-Day--selected)  {
+      background-color: @booked-color;
+    }
+    &--selected:not(.DayPicker-Day--outside) {
+      color: #FFF;
+      background-color: #4A90E2;
+    }
+  }
+
+  .DayPicker-WeekdaysRow {
+    background-color: @gray-lighter;
+    border-bottom: 1px solid white;
+    padding: 0.2rem;
+  }
+
+  .DayPicker-Weekdays {
+    text-transform: uppercase;
+    color: @gray;
+    font-weight: 600;
+    abbr {
+      border: none;
+      cursor: default;
+      vertical-align: text-top;
+    }
+  }
+
+  .calendar-legend {
+    display: flex;
+    justify-content: space-between;
+    span {
+      padding: 10px;
+    }
+    span::before {
+      content: '';
+      display: inline-block;
+      vertical-align: middle;
+      width: 16px;
+      height: 16px;
+      margin: 6px;
+      border-radius: 50%;
+      border: 2px solid black;
+    }
+    span.free::before {
+      background-color: @available-color;
+      border-color: darken(@available-color, 20);
+    }
+    span.busy::before {
+      background-color: @busy-color;
+      border-color: darken(@busy-color, 20);
+    }
+    span.booked::before {
+      background-color: @booked-color;
+      border-color: darken(@booked-color, 20);
+    }
+
+  }
+}
