+++ app/shared/shared.less
@@ -12,6 +12,7 @@
 @import './notifications/notifications';
 @import './reservation-confirmation/confirm-reservation-modal';
 @import './resource-icons/resource-icons';
+@import './resource-calendar/resource-calendar';
 @import './resource-list/resource-list';
 @import './resource-type-filter/resource-type-filter';
 @import './test-site-message/test-site-message';
